Update EventHandler.php
You guys wrote a really good code here. For this to stay up for 3 years. However, there are a few things I think can be adjusted. 

If the determineNotification() method is unable to find a matching notification class, it throws a ‘NotificationCouldNotBeSent’ exception. I think this exception is not caught or handled anywhere, so it would be better to add a try-catch block to handle this exception. Also, it’s clear the determineNotification() method should have a return type declaration to ensure that it returns an instance of the Notification class. You might want to rename some method names like determineNotifiable() and determineNotification() as they do not seem very clear.