Update installation-and-setup.md
Not sure if this is the best spot for it but I had "command not found" issues when running backup:run from my app while using Laravel Herd on macOS.

Happy for anyone to make/suggest changes or for it to go somewhere else entirely.

Perhaps there's even a better way to get around this issue? Open to suggestions.