Use MariaDb dumper
The spatie/laravel-backup package currently uses MySQL dumper for all MySQL-compatible databases, including MariaDB.

However, MariaDB has deprecated the mysqldump utility in favor of mariadb-dump as documented in their official documentation.

Recently, a dedicated MariaDB dumper was added to the spatie/db-dumper package in PR #220, which properly uses the mariadb-dump utility.

This PR updates this package to utilize the new dumper when a MariaDB connection is required.

on-behalf-of: @e-solutions-GmbH info@esolutions.de